fix byte-vs-string issue in interactive assembler
The monitor runs with stdin opened in binary mode, so read() returns
bytes. This resulted in errors from  nice65.utils.console.line_input,
which attempts to append a byte to a string.

This resolves the problem by ensuring that the getchar methods always
return a string (by calling .decode() when returning a character).